Use centerRegistrationPoint with New Game button
The button was not correctly aligned before. This happened due
to the fact that centerRegistrationPoint was called before moveTo.
The button is a child of a container, its registration point is centered,
which increases the width of the outer container in negative x-direction
and therefore width / 2 yields an undesired value.

If we move the button to its position before centering its registration
point we get the desired behaviour.